\name{spTDyn-package}
\alias{spTDyn-package}
\alias{spTDyn}
\docType{package}
\title{
Spatially varying and spatio-temporal dynamic linear models
}
\description{
This package uses different hierarchical Bayesian spatio-temporal modelling strategies, namely: \cr
(1) Spatially varying coefficient process models, \cr
(2) Temporally varying coefficient process models, also known as the spatio-temporal dynamic linear models.
}
\details{
\tabular{ll}{
Package: \tab spTDyn\cr
Type: \tab Package\cr
}
The back-end code of this package is built under c language.\cr
Main functions used: \cr
\code{> \link{GibbsDyn}} \cr
\code{> \link{predict.spT}} \cr
}
\author{
K.S. Bakar \cr
Maintainer: K.S. Bakar <shuvo.bakar@gmail.com>
}
\references{
 Bakar, K. S., Kokic, P. and Jin, H. (2015). A spatio-dynamic model for assessing frost risk in south-eastern Australia. Journal of the Royal Statistical Society, Series C. DOI: 10.1111/rssc.12103 \cr
 Bakar, K. S., Kokic, P. and Jin, H. (2015). Hierarchical spatially varying coefficient and temporal dynamic process models using spTDyn. Journal of Statistical Computation and Simulation. DOI:10.1080/00949655.2015.1038267
}
\keyword{ package }
\seealso{
Packages 'spTimer'; 'forecast'; 'spBayes'; 'maps'; 'MBA'; 'coda'; website: \code{http://www.r-project.org/}.
}
%\examples{
%~~ simple examples of the most important functions ~~
%}