\name{spTimerDyn-internal}
\title{Service functions and some undocumented functions for the spTimer library}
\alias{.First.lib}
\alias{.onLoad}
\alias{.onAttach}
\alias{stat.sum}
\alias{Formula.matrix}
\alias{Formula.coords}
\alias{Gamm}
\alias{Norm}
\alias{Unif}
\alias{PMCC}
\alias{fnc.time}
\alias{initials.checking.ar}
\alias{initials.checking.gp}
\alias{initials.checking.gpp}
\alias{print.spTD}
\alias{print.spTpred}
\alias{print.spTprd}
\alias{print.spTfore}
\alias{priors.checking.ar}
\alias{priors.checking.gp}
\alias{priors.checking.gpp}
\alias{spAR.Gibbs}
\alias{spAR.MCMC.Pred}
\alias{spAR.forecast}
\alias{spAR.prediction}
\alias{spGP.Gibbs}
\alias{spGP.MCMC.Pred}
\alias{spGP.forecast}
\alias{spGP.prediction}
\alias{spGPP.Gibbs}
\alias{spGPP.MCMC.Pred}
\alias{spGPP.forecast}
\alias{spGPP.prediction}
\alias{spT.Gibbs}
\alias{spT.priors}
\alias{spT.initials}
\alias{spT.decay}
\alias{spT.fit.pred}
\alias{spT.hit.false}
\alias{spT.forecast}
\alias{spT.prediction}
\alias{spT.Summary.Stat}
\alias{spT.MCMC.plot}
\alias{spT.MCMC.stat}
\alias{as.mcmc.spTD}
\alias{MCMC.plot.obj}
\alias{coef.spTD}
\alias{model.frame.spT}
\alias{residuals.spTD}
\alias{formula.spT}
\alias{terms.spT}
\alias{model.matrix.spT}
\alias{sp.dimname.fnc}
\alias{tp.dimname.fnc}
\alias{spT.keep.morethan.dist}
\alias{spT.grid.coords}
\alias{spT.check.locations}
\alias{spT.segment.plot}
\alias{spT.subset}
\alias{spT.geodist}
\alias{spT.pCOVER}
\alias{spT.validation}
\alias{confint.spTD}
\alias{fitted.spTD}
\alias{spT.check.sites.inside}
\alias{spT.geo.dist}
\alias{spT.geo_dist}
\alias{spT.time}
\alias{as.forecast.object}
\alias{NYdata}
\alias{AUSdata}
\alias{truncated.fnc}
\alias{reverse.truncated.fnc}
\alias{prob.below.threshold}
%\alias{NYgrid}
%\alias{imputation.z}
\description{Internal spTDyn functions}
\keyword{internal}
